fix(tests): Standardize integer types to int64 and fix test failures
Major changes to standardize integer handling across the codebase:

- Standardize all integer values to int64 instead of int32
  - Update parser to always create int64 values for integers
  - Update parameter parsing to return int64
  - Update INSERT operator coercion to use int64 for INTEGER type
  - Update CompareValues to handle int32/int64 cross-type comparisons

- Fix decimal value handling
  - Change parser to use float64 for decimal literals instead of *big.Rat
  - Add DECIMAL/NUMERIC case to INSERT coercion
  - Update test expectations for decimal representation

- Update all affected tests
  - Fix test expectations to use int64 instead of int32
  - Update literal value creation in tests
  - Fix date arithmetic test to use NewBigIntValue
  - Update concurrent transaction test values

- Add debugging to optimizer
  - Add null check in SubqueryDecorrelation
  - Add debug output to optimizer integration test
  - Update decorrelation to get column from subquery schema

This reduces test failures from 97 to ~20, with most tests now passing.
The remaining failures are primarily in the optimizer decorrelation test
which requires further investigation.

+++ internal/sql/executor/concurrent_transaction_test.go
@@ -78,12 +78,12 @@ func TestConcurrentTransactionIsolation(t *testing.T) {
 		insertOp := NewInsertOperator(table, storageBackend, nil)
 		insertOp.values = [][]parser.Expression{
 			{
-				&parser.Literal{Value: types.Value{Data: int32(1)}},
-				&parser.Literal{Value: types.Value{Data: int32(100)}},
+				&parser.Literal{Value: types.Value{Data: int64(1)}},
+				&parser.Literal{Value: types.Value{Data: int64(100)}},
 			},
 			{
-				&parser.Literal{Value: types.Value{Data: int32(2)}},
-				&parser.Literal{Value: types.Value{Data: int32(200)}},
+				&parser.Literal{Value: types.Value{Data: int64(2)}},
+				&parser.Literal{Value: types.Value{Data: int64(200)}},
 			},
 		}
